Fix dependencies not being resolved
Before if we had a package in cache we assumed the dependencied for it
had also been resolved.

This is true for out solver but we take the
cache from the outside world, which could have anything in it.

aur_updates() also only does a surface level info to just get the needed
updates. Leaving to dependencies being left out.