Fix target flag being set on dependencies
Within resolve_aur_pkg(), the AurPackage structs added to the build
action list are only for the dependent packages. However, the first call
to resolve_aur_pkg() can be called with target=true, and this value was
being used to set the target flag for each of these AurPackages. This is
incorrect and should simply be set to false for all dependencies.

Any recursive dependencies were being set correctly, as
resolve_aur_pkg() was explicitly called with target=false.

Since the primary target package is added to the build action list
separately in resolve(), this is the only place were the AurPackage
needs target=true, and is handled independently of resolve_aur_pkg().

To fix, remove the target logic from resolve_aur_pkg() entirely. Also
rename the function to better indicate that it is only resolving
dependencies.

Also rename some confusing aliasing of the pkg parameter.

Finally, add a test to catch the original bug.