Try to order targets by deps even when satisfied
When trying to install sway and wlroots, if wlroots is already
installed, sway does not pull wlroots in as a dep and instead wlroots
then gets pulled in second as a target.

While there's nothing technically wrong with this, because AUR packages
are built from source, wlroots could have a new .so name which would cause
sway to then break even though they were installed together.

This commit makes it so if a dep is satisfied by a target then we ignore
the fact that it may already be satisfied and pull it in anyway.