Resolve pkgbuild deps between split packages
When resolving deps, if the package is split and package A depends on
package B. if package B happened to also exist in the repos then we
would pick the dep from there instead of the pkgbuild.

This leads to issues when trying to resolve A and B as we end up with
B getting pulled from the repos and we end up with duplicate packages.